path building works nice, but has some bugs (when you use it too fast or when you press right mouse button)

+++ Assets/Scripts/Cards/Card.cs
@@ -10,7 +10,7 @@ namespace Cards
         [SerializeField] protected Sprite FaceSprite;
         protected ulong CaptorId;
         private Unit _currentUnit;
-        [SerializeField] public ulong id;
+        [SerializeField] public int id;
 
         //get/set card visibility (also calls method for flipping card on visibility change)
         public bool IsVisible
